+++ vsc-script/.vsc-script/util/with__import.ts
@@ -1,0 +1,8 @@
+
+export const test = () => {
+   return "hej"
+}
+export const test2 = () => {
+   return "hej2"
+}
+export const test3 = "WHAT!!"