+++ vsc-script/.vsc-script/with__import.vsc-script.error-log.js
@@ -1,0 +1,110 @@
+
+// ------------------ Error log from script -------------- //
+// type: 104.2: Error in vsc-Script trying to transpile the loaded module. Please report it to https://github.com/alfnielsen/vsc-base/issues and include the error-log
+// time: 2019-07-05T04:33:36.894Z
+// selectedScript.path: /Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/with__import.vsc-script.ts
+// path: [object Object]
+// error:
+// Error handler did not find the extension.js!
+var info = {
+   "isError": true,
+   "type": "",
+   "stack": "Error: ENOENT: no such file or directory, open 'vsc-base'",
+   "message": "ENOENT: no such file or directory, open 'vsc-base'"
+}
+// ts transpiled js code:
+/Users/alfnielsen / Dropbox / Coding / vsc - base / vsc - script / .vsc - script / with__import.vsc - script.ts;
+
+
+
+// ------------------ Error log from script -------------- //
+// type: 104.2: Error in vsc-Script trying to transpile the loaded module. Please report it to https://github.com/alfnielsen/vsc-base/issues and include the error-log
+// time: 2019-07-05T04:36:04.945Z
+// selectedScript.path: /Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/with__import.vsc-script.ts
+// path: [object Object]
+// error:
+// Error handler did not find the extension.js!
+var info = {
+   "isError": true,
+   "type": "",
+   "stack": "Error: ENOTDIR: not a directory, open 'Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/with__import.vsc-script.ts/vsc-base'",
+   "message": "ENOTDIR: not a directory, open 'Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/with__import.vsc-script.ts/vsc-base'"
+}
+// ts transpiled js code:
+/Users/alfnielsen / Dropbox / Coding / vsc - base / vsc - script / .vsc - script / with__import.vsc - script.ts;
+
+
+
+// ------------------ Error log from script -------------- //
+// type: 104.2: Error in vsc-Script trying to transpile the loaded module. Please report it to https://github.com/alfnielsen/vsc-base/issues and include the error-log
+// time: 2019-07-05T04:41:24.543Z
+// selectedScript.path: /Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/with__import.vsc-script.ts
+// path: [object Object]
+// error:
+// Error handler did not find the extension.js!
+var info = {
+   "isError": true,
+   "type": "",
+   "stack": "Error: ENOENT: no such file or directory, open 'Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/./with__import'",
+   "message": "ENOENT: no such file or directory, open 'Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/./with__import'"
+}
+// ts transpiled js code:
+/Users/alfnielsen / Dropbox / Coding / vsc - base / vsc - script / .vsc - script / with__import.vsc - script.ts;
+
+
+
+// ------------------ Error log from script -------------- //
+// type: 104.1: Error in vsc-Script trying to transpile the loaded module. This error is properly incorrect formatting of the module file. Open the script file in vscode and ensure that ts cont find any errors.
+// time: 2019-07-05T04:45:42.648Z
+// selectedScript.path: /Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/with__import.vsc-script.ts
+// path: [object Object]
+// error:
+// Error handler find first error in extension.js: 
+var info = {
+   "isError": true,
+   "type": "",
+   "stack": "Error: ReferenceError: __vsc__import__exports is not defined\n\tat loadTsModule_Eval (/Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/dist/extension.js:136718:15)\n\tat Object.<anonymous> (/Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/dist/extension.js:136691:20)\n\tat Generator.next (<anonymous>)\n\tat fulfilled (/Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/dist/extension.js:136551:58)",
+   "message": "ReferenceError: __vsc__import__exports is not defined"
+}
+// ts transpiled js code:
+__vsc__import__exports = importExports;
+_exports = (function(){var exports = {};
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.test = () => {
+    return "hej";
+};
+exports.test2 = () => {
+    return "hej2";
+};
+
+return exports})()
+
+
+// ------------------ Error log from script -------------- //
+// type: 104.1: Error in vsc-Script trying to transpile the loaded module. This error is properly incorrect formatting of the module file. Open the script file in vscode and ensure that ts cont find any errors.
+// time: 2019-07-05T04:47:11.252Z
+// selectedScript.path: /Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/.vsc-script/with__import.vsc-script.ts
+// path: [object Object]
+// error:
+// Error handler find first error in extension.js: 
+var info = {
+   "isError": true,
+   "type": "",
+   "stack": "Error: ReferenceError: __vsc__import__exports is not defined\n\tat loadTsModule_Eval (/Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/dist/extension.js:136718:15)\n\tat Object.<anonymous> (/Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/dist/extension.js:136691:20)\n\tat Generator.next (<anonymous>)\n\tat fulfilled (/Users/alfnielsen/Dropbox/Coding/vsc-base/vsc-script/dist/extension.js:136551:58)",
+   "message": "ReferenceError: __vsc__import__exports is not defined"
+}
+// ts transpiled js code:
+__vsc__import__exports = importExports;
+_exports = (function(){var exports = {};
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.test = () => {
+    return "hej";
+};
+exports.test2 = () => {
+    return "hej2";
+};
+
+return exports})()
+